requirements.txt
discord >= 2.3.0
requests
asyncpg
gitpython
dnspython
plotly
pillow